requirements.txt
alembic==1.17.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
bcrypt==4.3.0
beautifulsoup4==4.14.2
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.119.1
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.11.1
kombu==5.5.4
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.3.4
openai==2.6.0
outcome==1.3.0.post0
packaging==25.0
passlib==1.7.4
pgvector==0.4.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.1
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
requests==2.32.5
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
rsa==4.9.1
selenium==4.27.1
sentry-sdk==2.42.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.48.0
tqdm==4.67.1
trio==0.31.0
trio-websocket==0.12.2
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==15.0.1
wsproto==1.2.0
